+++ CHANGELOG.md
@@ -58,6 +58,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   or a direct reference to a `<code-sample-project>` element (which would allow
   the elements to live in different scopes).
 
+- The caret is now only displayed when an editor is on focus (previously it was
+  always displayed).
+
 ### Added
 
 - Add syntax highlighting for TypeScript files.